Produce a density plot of iPAS permutation scores for a given pathway

This function returns a plot of the probability density of the null distribution of iPAS scores for permutation signatures for a given pathway. The iPAS score for the (non-permuted) input signature is shown by a red bar.

iPAS_density(
  res,
  experiment,
  score_type = c("z_score", "score"),
  path,
  fill = "light blue"
)

Arguments

res
output object from the iPAS_enrich function

experiment
the sample number or name of the sample from which to graph results.

score_type
the type of iPAS score to graph. "score" will show raw iPAS scores and "z_score" will show normalized iPAS scores.

path
the KEGG pathway ID of the pathway to graph (e.g. "hsa04150" for the mTOR pathway)

fill
the color to fill in the density plot, "light blue" by default.